Create a helper function for formatting dates
“ember g util date”

Utils are where generic javascript goes.  Mike makes the case for putting general javascript here and bringing it into components.  This makes unti testing very easy/simple, and sets it up to be reusable because it’s just plain javascript.

“ember g helper format-timestamp”

helper function becomes available in templates.  Brings in one of the functions from the date utils.