Create a component action for login
This commit creates an action, which is a decorator to bind the value of “this” to the component context.  In this example, without the action “this” is trying to find a method on the DOM event, not the js component.

Adding the action resolves this to the component so we can use the method we defined in the event handler.

also creates an integration test using a regex Mike Norht defined.