Add component state & opt in to UI reactivity
This commit adds a component method to update the state of the compont when a user switches the selected user in the dropdown, and subscribes the UI to changes in this tracked property.

Difference between Ember classic and Octane — ember classic promoted 2-way data binding, Octane promotes 1-way data binding so we have to explicitly declare the reactive property.