Add logout to auth service, use and test it
This commit adds the logout functionality in the auth service, then sets up the team sidebar component to use the logout method.

Logout is an action because we use it directly in the team-sidebar template {{on “click” this.auth.logout}}.

To get access to the service we create a team-sidebar.js that simply injects the auth service.