Make UI in team-sidebar dynamic from route model
Our team model gets the teamId as a model because of the router query parameter (setup in router.js).

We use this teamId that the route gets in the team route model hook, compare against the array of all teams, to make the team object accessible to the team template.

We then use a property on that object that the team template now has to dynamically render the team name when a user switches the route.  Cool!