Dynamically get channels data, use paramsFor to get parent route
This commit gets and renders dynamic data for the channels.

Note the usage of paramsFor to dyanically generate the route needed for the api call — we go “up” in the route stack to pull data (teamId) from a parent route.

M.N. advice:  “always go up, not sideways or down” in terms of accessing data from parent/sibling/child routes.