Create container component logic, pass to child
This commit sets up the actions to create a new message in the database.  The api handler is in the container component, which gets yielded through a hash.

This then gets passed to channel-footer (which is a mostly presentational component except for some minor internal component state).  It’s expecting this action to be passed as an arg, which we do.

Also some minor UI state on the button logic using the getter for whether or not the footer button should be disabled.

The thinking here is that you want to separete pretty clearly which components are going to be smart and stateful (our chat-contianer) and which ones are mostly presentational (our channel-footer).